requirements.txt
ultralytics
numpy
pandas
matplotlib
boto3
psycopg2-binary
pytest
pytest-cov
